# src/DocHub.Api/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DocHub.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/content-styles": {
      "get": {
        "tags": [
          "Content styles"
        ],
        "summary": "Styles for the editor's style dropdown, optionally of one kind.",
        "operationId": "ListContentStyles",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentStyleKind"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentStyleResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Content styles"
        ],
        "summary": "Admin: creates a style (styleId unique; properties validated against the style property schema).",
        "operationId": "CreateContentStyle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentStyle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentStyleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/content-styles/stylesheet.css": {
      "get": {
        "tags": [
          "Content styles"
        ],
        "summary": "CSS generated from the catalog (.ds-style-Heading1 { … }), revalidated with ETag.",
        "operationId": "GetContentStylesheet",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/css": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified"
          }
        }
      }
    },
    "/api/content-styles/{id}": {
      "get": {
        "tags": [
          "Content styles"
        ],
        "summary": "A style; usageCount = contents using it.",
        "operationId": "GetContentStyle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentStyleResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Content styles"
        ],
        "summary": "Admin: updates a style (styleId and kind are fixed; rowVersion required); isActive = false deactivates it.",
        "operationId": "UpdateContentStyle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentStyle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentStyleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Content styles"
        ],
        "summary": "Admin: deletes an unused custom style (?rowVersion=base64); built-in → 409 built-in-style, used → 409 in-use.",
        "operationId": "DeleteContentStyle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "rowVersion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ContentStyleKind": {
        "enum": [
          "Paragraph",
          "Character",
          "Table"
        ]
      },
      "ContentStyleResponse": {
        "required": [
          "id",
          "styleId",
          "name",
          "kind",
          "basedOnStyleId",
          "properties",
          "isBuiltIn",
          "isActive",
          "rowVersion",
          "usageCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "styleId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ContentStyleKind"
          },
          "basedOnStyleId": {
            "type": [
              "null",
              "string"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "isBuiltIn": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          },
          "usageCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CreateContentStyle": {
        "required": [
          "styleId",
          "kind",
          "name"
        ],
        "type": "object",
        "properties": {
          "styleId": {
            "pattern": "^[A-Za-z][A-Za-z0-9]{0,49}$",
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContentStyleKind"
              }
            ]
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "type": [
              "null",
              "string"
            ]
          },
          "basedOnStyleId": {
            "maxLength": 50,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/JsonElement"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "JsonElement": {},
...
      "UpdateContentStyle": {
        "required": [
          "rowVersion",
          "name"
        ],
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "rowVersion": {
            "type": [
              "null",
              "string"
            ],
            "format": "byte"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "type": [
              "null",
              "string"
            ]
          },
          "basedOnStyleId": {
            "maxLength": 50,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/JsonElement"
          }
        }
      },
...
    },
    "securitySchemes": {
      "TestMode": {
        "type": "apiKey",
        "description": "Test mode: id of a seeded, active user (e.g. 2 = alice).",
        "name": "X-User-Id",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "TestMode": []
    }
  ],
...
}